Fixed conflicting file names, introduced working template for changing scenes
